refactor(analysers): migrate to utils
