+++ pdf_craft/analysers/correction/corrector.py
@@ -7,8 +7,7 @@ from xml.etree.ElementTree import Element
 
 from ...llm import LLM
 from ...xml import encode_friendly
-from ..context import Context
-from ..partition import Partition
+from ..utils import Context, Partition
 from ..sequence import read_paragraphs, Paragraph, ParagraphType
 from .common import State, Phase
 from .repeater import repeat_correct
