+++ pdf_craft/analysers/utils/context.py
