+++ pdf_craft/analysers/utils/others.py
