+++ pdf_craft/analysers/utils/partition.py
